파일 처리 로직 변경.
